requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.2.1
feedparser==6.0.10
frozendict==2.3.8
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
multitasking==0.0.11
numpy==1.25.2
pandas==2.0.3
parse==1.19.1
psycopg2==2.9.6
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.27
zipp==3.16.2
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
cssselect==1.2.0
fake-useragent==1.2.1
feedparser==6.0.10
frozendict==2.3.8
greenlet==2.0.2
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
lxml==4.9.3
multitasking==0.0.11
numpy==1.25.2
pandas==2.0.3
parse==1.19.1
psycopg2==2.9.6
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-html==0.10.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
yahoo-fin==0.8.9.1
yfinance==0.2.27
zipp==3.16.2
